AEIM-1076 - Split beats into elastic, elastic::metricbeat, and elastic::filebeat
