+++ manifests/profile/elastic.pp
@@ -2,10 +2,7 @@
 # All Rights Reserved. Licensed according to the terms of the Revised
 # BSD License. See LICENSE.txt for details.
 
-# beats
-#
-# Install and configure metricbeat, and ensure that its service is
-# running.
+# Elastic Beats
 #
 # @param logstash_auth_cert If (and only if) defined as a 'puppet://'
 #   path, logstash will use the source file as its ssl certificate
@@ -14,55 +11,12 @@
 # @param period System module period in seconds
 #
 # @example
-#   include nebula::profile::beats
-class nebula::profile::beats (
+#   include nebula::profile::elastic
+class nebula::profile::elastic (
   String  $logstash_auth_cert,
   Array   $logstash_hosts,
   Integer $period,
 ) {
-  service { 'metricbeat':
-    ensure     => 'running',
-    enable     => true,
-    hasrestart => true,
-    require    => File['/etc/metricbeat/metricbeat.yml'],
-  }
-
-  service { 'filebeat':
-    ensure     => 'running',
-    enable     => true,
-    hasrestart => true,
-    require    => File['/etc/filebeat/filebeat.yml'],
-  }
-
-  file { '/etc/metricbeat/metricbeat.yml':
-    ensure  => 'present',
-    mode    => '0644',
-    content => template('nebula/profile/beats/metricbeat.yml.erb'),
-    require => Package['metricbeat'],
-    notify  => Service['metricbeat'],
-  }
-
-  file { '/etc/filebeat/filebeat.yml':
-    ensure  => 'present',
-    mode    => '0644',
-    content => template('nebula/profile/beats/filebeat.yml.erb'),
-    require => Package['filebeat'],
-    notify  => Service['filebeat'],
-  }
-
-  file { '/etc/filebeat/prospectors':
-    ensure  => 'directory',
-    require => Package['filebeat'],
-  }
-
-  package { 'metricbeat':
-    require => Apt::Source['elastic.co'],
-  }
-
-  package { 'filebeat':
-    require => Apt::Source['elastic.co'],
-  }
-
   apt::source { 'elastic.co':
     comment  => 'Elastic.co apt source for beats and elastic search',
     location => 'https://artifacts.elastic.co/packages/5.x/apt',
