+++ templates/profile/elastic/filebeat.yml.erb
