+++ templates/profile/elastic/metricbeat.yml.erb
